fix: devices.entityId INT→BIGINT—Rust+设备entityId可超21亿(如Smart Alarm 2809707316),原INT致配对INSERT报Out-of-range、设备存不进库、网页不展示也收不到触发

+++ backend/sql/init.sql
@@ -79,7 +79,7 @@ CREATE TABLE IF NOT EXISTS `devices` (
   `id` VARCHAR(36) NOT NULL,
   `serverId` VARCHAR(36) NOT NULL,
   `userId` VARCHAR(36) NOT NULL,
-  `entityId` INT NOT NULL,
+  `entityId` BIGINT NOT NULL,
   `name` VARCHAR(100) NOT NULL,
   `type` ENUM('SWITCH', 'ALARM', 'STORAGE') NOT NULL,
   `command` VARCHAR(50) NULL,
